Adding a date in submission to know when was the last replay & display info in submissions page.
Adding a date in submission to know when was the last replay & display info in submissions page.

[tests/filesystems] Add filesystem provider tests

The tests are applied on the LocalFSProvider.

[ci] Add pytest run + coverage report

The coverage reports are saved in GitHub artefacts for offline analysis.
Currently, the badge is also stored in artefacts but it is not
retrievable by the README.

[tests/filesystems] Make the tests modular with fixture parameter

[ci] Revert pip cache storage

move replay date set to submission_manager

move replay date set to submission_manager