[frontend] Adding timeZone to last datetimepickers
Had a problem when retrieving taskid in accessibility.html. It is passed to the button on the opening of the modal. However they are two events triggered by that opening. The one where defaultDate are given to the datetimepicker triggers first. And the change in value triggers another event using taskid. Therefor taskid wass accessed before being passed as data.