[frontend/accessibility] Applying timezone to task accessibility edit
Because the modal is only one HTML code block we update with new data, datetime pickers need to be destroyed are recreated with the new defaultDate. This is the only option found to apply the timezone. Changing the date with datetimepicker('date', ...) does not take timezone into consideration.

Recreating the datetime picker triggers the change event listeners. Added a 'ignore-change' value to pickers to avoid updating the dispenser_config on modal feed.