[frontend/user_manager] Code indentation not in DB by default + removing ancient DB update when connecting user
Not setting the code indentation in DB if the user has not changed it from default allows the gain of space in the database. A database update was present in the user connection because the idea in the past was to have the preferences changed in session and in DB through the same method. This was aaplied for the username change but was not done later for other preferences. A DB update was added later in profile.py despite beeing already present in connect_user().